fix: record Apple overhead gate constants as exact decimals
make bench-apple-overhead serialized kParityMaxAbs (Swift Float 5e-4)
into the receipt as the f32 bit pattern 0.0005000000237487257, so
apple_overhead_receipt failed its exact comparison against 0.0005 even
on an all-green Machine C run.

- Swift: kParityMaxAbs is now Double, so gates.parity_max_abs and each
  case's max_abs_gate round-trip as 0.0005. The per-case parity check
  compares Double(maxAbs) against the same Double gate, matching the
  Rust validator exactly at the boundary.
- Rust: recorded gate constants are accepted as either the exact f64
  value or precisely its f32 rounding (for receipts produced by the
  Float-serializing harness). Case/repeat gates still use the exact
  f64 budgets; no performance budget or parity gate is relaxed.

+++ swift/Sources/BenchAppleOverhead/BenchAppleOverhead.swift
@@ -43,7 +43,10 @@ import MLXLMCommon
 import Metal
 import MlxEngine
 
-let kParityMaxAbs: Float = 5e-4
+// Double, not Float: the gate constant is recorded in the receipt, and the
+// f32 bit pattern of 5e-4 serializes as 0.0005000000237… instead of 0.0005,
+// which the Rust receipt validator rejects.
+let kParityMaxAbs: Double = 5e-4
 let kParityMaxRmse = 1e-4
 let kP50OverheadLimit = 1.05
 let kThroughputFloor = 0.95
@@ -369,7 +372,7 @@ struct BenchAppleOverhead: AsyncParsableCommand {
                 sqSum += Double(diff) * Double(diff)
             }
             let rmse = (sqSum / Double(abiOut.count)).squareRoot()
-            let parityOk = maxAbs <= kParityMaxAbs && rmse <= kParityMaxRmse
+            let parityOk = Double(maxAbs) <= kParityMaxAbs && rmse <= kParityMaxRmse
             if !parityOk { allPass = false }
 
             // Steady-state ABI allocation counter on one post-warmup request.
@@ -436,7 +439,7 @@ struct BenchAppleOverhead: AsyncParsableCommand {
                 "parity": [
                     "max_abs": Double(maxAbs),
                     "rmse": rmse,
-                    "max_abs_gate": Double(kParityMaxAbs),
+                    "max_abs_gate": kParityMaxAbs,
                     "rmse_gate": kParityMaxRmse,
                     "pass": parityOk,
                 ],
@@ -493,7 +496,7 @@ struct BenchAppleOverhead: AsyncParsableCommand {
             "gates": [
                 "abi_p50_over_direct_p50_max": kP50OverheadLimit,
                 "abi_throughput_over_direct_min": kThroughputFloor,
-                "parity_max_abs": Double(kParityMaxAbs),
+                "parity_max_abs": kParityMaxAbs,
                 "parity_rmse": kParityMaxRmse,
                 "p99_min_samples": kP99MinSamples,
             ],
